Consolidate duplicated category emoji maps into shared modules
Extract category-to-emoji mappings from 5 inline copies (3 frontend,
2 server-side) into two canonical modules — one per side of the
client/server boundary. Fixes stale emojis in src/main.js and removes
dead PRManager.CATEGORY_EMOJI_MAP getter.
