+++ public/local.html
@@ -497,6 +497,9 @@
     <!-- Tier icons utility -->
     <script src="/js/utils/tier-icons.js"></script>
 
+    <!-- Category emoji mapping -->
+    <script src="/js/utils/category-emoji.js"></script>
+
     <!-- Suggestion UI utility -->
     <script src="/js/utils/suggestion-ui.js"></script>
