+++ public/pr.html
@@ -320,6 +320,9 @@
     <!-- Tier icons utility -->
     <script src="/js/utils/tier-icons.js"></script>
 
+    <!-- Category emoji mapping -->
+    <script src="/js/utils/category-emoji.js"></script>
+
     <!-- Suggestion UI utility -->
     <script src="/js/utils/suggestion-ui.js"></script>
